api/pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "bcrypt<4",
    "fastapi>=0.115.12",
    "gunicorn>=25.3.0",
    "passlib>=1.7.4",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "pydantic[email]>=2.11.4",
    "pyjwt>=2.11.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "uvicorn>=0.34.2",
    "yoyo-migrations>=9.0.0",
]
